9-web-app-integration/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.4
python-dotenv==1.0.1
pydantic==2.10.4

python-multipart==0.0.5
pdfplumber
